Speedup .digest, .hexdigest and .base64digest class methods
Allocating the hasher in HEAP memory (+ maybe the whole class and
method calls) are destroying the performance of BLAKE3 so much that it
was slightly slower than Digest::SHA256!

This patch overrides the methods from Digest::ClassMethods to allocate
the hasher on the stack and directly call the C functions, so BLAKE3 is
now 3 to 4 times faster than Digest::SHA256.

I tried to wrap the hasher into a struct, and share the implementation
between the class and the class methods, but performance got destroyed
again (only 1.5 times faster than Digest::SHA256). I have no explanation
for this huge performance drop.